+++ eras/allegra/impl/src/Cardano/Ledger/Allegra/PParams.hs
@@ -9,7 +9,6 @@ import Cardano.Ledger.Crypto
 import Cardano.Ledger.Shelley.Governance
 import Cardano.Ledger.Shelley.PParams
 import Data.Coerce
-import qualified Data.Map.Strict as Map
 import Lens.Micro
 
 instance Crypto c => EraPParams (AllegraEra c) where
@@ -58,5 +57,3 @@ instance Crypto c => EraGov (AllegraEra c) where
   prevPParamsGovStateL = prevPParamsShelleyGovStateL
 
   obligationGovState = const mempty
-
-  getDRepDistr = const Map.empty